requirements.txt
polars==1.5.0
pandas==2.2.2
psycopg2==2.9.9
xlrd==2.0.1
arrow==1.3.0
loguru==0.7.2
apache-airflow-providers-postgres==5.8.0
